rename files in default-theme to use current schema
